pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chirpscan-object-detection"
version = "0.1.0"
description = "YOLO-based object detection on spectrograms for bioacoustic analysis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ChirpScan Team"}
]

dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "ultralytics>=8.0.200",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pillow>=10.0.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "ipython>=8.12.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]

experiment = [
    "wandb>=0.16.0",
    "tensorboard>=2.15.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N"]
ignore = []
target-version = "py311"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"
